pyproject.toml
[tool.poetry]
name = "people-detection"
version = "0.1.0"
description = ""
authors = ["Adilson Pavan <adilson.pavan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
opencv-contrib-python = "^4.3.0"
python-dotenv = "^0.14.0"
numpy = "^1.19.0"
python-telegram-bot = "^12.8"
imutils = "^0.5.3"

[tool.poetry.dev-dependencies]
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
